refactor: Move out new entities into `wasm` module
