+++ tests/src/common/dispatcher.rs
@@ -6,7 +6,7 @@ use rewards_dispatcher::{
 use sdk::cosmwasm_std::Addr;
 
 use super::{
-    test_case::app::{App, Wasm as WasmTrait},
+    test_case::{app::App, wasm::Wasm as WasmTrait},
     CwContractWrapper, ADMIN,
 };
