+++ tests/src/common/lpp.rs
@@ -13,7 +13,7 @@ use sdk::{
 };
 
 use super::{
-    test_case::app::{App, Wasm as WasmTrait},
+    test_case::{app::App, wasm::Wasm as WasmTrait},
     CwContractWrapper, ADMIN,
 };
